assignments/autogen-dev-framework/requirements.txt
autogen-agentchat~=0.2
flaml[automl]
python-dotenv
openai
typing-extensions
requests
pytest